html/semantics/embedded-content/the-img-element/update-the-image-data/set-src-idl-same-value-loaded-not-cacheable.html

<!doctype html>
<title>img set .src IDL attribute to same value (loaded, not cacheable)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/common.js"></script>
<div id=log></div>
<script>
var url_settings = {
  cached: "no",
  animated: "no",
  delay: "no",
  broken: "no",
};
var expected_requests = '1';
var expected_props = {
  width: 1,
  height: 1,
  complete: true,
};

async_test(function() {
  var img = new Image();
  var url = make_url(url_settings);
  var url_put = url + 'put';
  var url_take = url + 'take';
  img.onload = this.step_func(function() {
    assert_equals(img.width, 1);
    assert_equals(img.complete, true);
    img.onload = this.step_func(function() {
      this.step(check_requests(url_take, expected_requests));
      img.onload = this.unreached_func("got third load event");
    });
    img.src = url_put;
  });
  img.onerror = this.unreached_func("got error event");
  img.src = url_put;
});
</script>
